refactor: Email 생성 시 자체 검증으로 Validator 책임 축소
- Validator 필드 개선 (불필요한 리플렉션 제거)
