+++ src/main/java/roomescape/common/domain/Email.java
@@ -8,7 +8,12 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 import lombok.experimental.FieldNameConstants;
+import lombok.extern.slf4j.Slf4j;
+import roomescape.common.validate.InvalidArgumentException;
+import roomescape.common.validate.ValidationType;
 import roomescape.common.validate.Validator;
+
+import java.util.regex.Pattern;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -19,6 +24,8 @@ import roomescape.common.validate.Validator;
 @Embeddable
 public class Email {
 
+    private static final Pattern EMAIL_PATTERN = Pattern.compile("^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+$");
+
     private String value;
 
     public static Email from(final String value) {
@@ -28,6 +35,16 @@ public class Email {
 
     private static void validate(final String value) {
         Validator.of(Email.class)
-                .validateEmailFormat(Fields.value, value, DomainTerm.EMAIL.label());
+                .validateNotBlank(Fields.value, value, DomainTerm.EMAIL.label());
+
+        if (EMAIL_PATTERN.matcher(value).matches()) {
+            return;
+        }
+
+        throw new InvalidArgumentException(
+                ValidationType.EMAIL_CHECK,
+                Email.class.getSimpleName(),
+                Fields.value,
+                DomainTerm.EMAIL.label());
     }
 }
