+++ src/main/java/roomescape/common/validate/InvalidArgumentException.java
@@ -2,14 +2,13 @@ package roomescape.common.validate;
 
 import org.springframework.http.HttpStatus;
 import roomescape.common.exception.base.BusinessException;
-import roomescape.common.validate.Validator.ValidationType;
 
-public class InvalidInputException extends BusinessException {
+public class InvalidArgumentException extends BusinessException {
 
-    public InvalidInputException(final ValidationType type,
-                                 final String className,
-                                 final String fieldName,
-                                 final String fieldDescription) {
+    public InvalidArgumentException(final ValidationType type,
+                                    final String className,
+                                    final String fieldName,
+                                    final String fieldDescription) {
         super(
                 buildLogMessage(type, className, fieldName),
                 buildUserMessage(fieldDescription)
