+++ src/test/java/roomescape/theme/domain/ThemeNameTest.java
@@ -2,7 +2,7 @@ package roomescape.theme.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.common.validate.InvalidInputException;
+import roomescape.common.validate.InvalidArgumentException;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -16,7 +16,7 @@ class ThemeNameTest {
         // when
         // then
         assertThatThrownBy(() -> ThemeName.from(null))
-                .isInstanceOf(InvalidInputException.class)
+                .isInstanceOf(InvalidArgumentException.class)
                 .hasMessage("Validation failed [while checking blank]: ThemeName.value");
     }
 
@@ -27,11 +27,11 @@ class ThemeNameTest {
         // then
         assertAll(() -> {
             assertThatThrownBy(() -> ThemeName.from(""))
-                    .isInstanceOf(InvalidInputException.class)
+                    .isInstanceOf(InvalidArgumentException.class)
                     .hasMessage("Validation failed [while checking blank]: ThemeName.value");
 
             assertThatThrownBy(() -> ThemeName.from(" "))
-                    .isInstanceOf(InvalidInputException.class)
+                    .isInstanceOf(InvalidArgumentException.class)
                     .hasMessage("Validation failed [while checking blank]: ThemeName.value");
         });
     }
